requirements.txt
chromadb
gradio
langchain
langchain-community
langchain-openai
openai
pypdf
python-dotenv
streamlit
tiktoken